change header's and chunk's unsigned long to unsigned int
unsigned long is 8 bytes on x64 unix while being 4 bytes on windows breaking the header and chunk structures

+++ main.cpp
@@ -8,20 +8,20 @@
 struct chunk_t
 {
 	char ID[4]; //"data" = 0x61746164
-	unsigned long size;  //Chunk data bytes
+	unsigned int size;  //Chunk data bytes
 };
 //Wav Header
 struct wav_header_t
 {
 	char chunkID[4]; //"RIFF" = 0x46464952
-	unsigned long chunkSize; //28 [+ sizeof(wExtraFormatBytes) + wExtraFormatBytes] + sum(sizeof(chunk.id) + sizeof(chunk.size) + chunk.size)
+	unsigned int chunkSize; //28 [+ sizeof(wExtraFormatBytes) + wExtraFormatBytes] + sum(sizeof(chunk.id) + sizeof(chunk.size) + chunk.size)
 	char format[4]; //"WAVE" = 0x45564157
 	char subchunk1ID[4]; //"fmt " = 0x20746D66
-	unsigned long subchunk1Size; //16 [+ sizeof(wExtraFormatBytes) + wExtraFormatBytes]
+	unsigned int subchunk1Size; //16 [+ sizeof(wExtraFormatBytes) + wExtraFormatBytes]
 	unsigned short audioFormat;
 	unsigned short numChannels;
-	unsigned long sampleRate;
-	unsigned long byteRate;
+	unsigned int sampleRate;
+	unsigned int byteRate;
 	unsigned short blockAlign;
 	unsigned short bitsPerSample;
 	//[WORD wExtraFormatBytes;]
